+++ aop/src/main/java/com/interface21/beans/factory/config/BeanPostProcessor.java
@@ -1,5 +1,7 @@
 package com.interface21.beans.factory.config;
 
 public interface BeanPostProcessor {
+    boolean requiresProcessing(Object bean);
+
     Object postInitialization(Object bean);
 }
