+++ di/build.gradle
@@ -16,6 +16,7 @@ repositories {
 
 dependencies {
     implementation project(':aop')
+    implementation project(':jdbc')
 
     implementation 'jakarta.annotation:jakarta.annotation-api:2.1.1'
     annotationProcessor 'jakarta.annotation:jakarta.annotation-api:2.1.1'
@@ -24,6 +25,7 @@ dependencies {
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.17.1'
     implementation 'ch.qos.logback:logback-classic:1.5.6'
     implementation 'org.apache.commons:commons-lang3:3.14.0'
+    implementation 'cglib:cglib:3.3.0'
 
     testImplementation 'org.assertj:assertj-core:3.26.0'
     testImplementation 'org.mockito:mockito-core:5.12.0'
