+++ di/src/main/java/com/interface21/beans/factory/ConfigurableListableBeanFactory.java
@@ -1,5 +1,9 @@
 package com.interface21.beans.factory;
 
+import com.interface21.beans.factory.config.BeanPostProcessor;
+
 public interface ConfigurableListableBeanFactory extends BeanFactory {
+    void addPostProcessor(BeanPostProcessor beanPostProcessor);
+
     void preInstantiateSingletons();
 }
